Update autoyxip_macos.sh
自动优选ip macOS版本 支持v6v4 DDNS

+++ 自动优选ip/autoyxip_macos.sh
@@ -1,10 +1,24 @@
 #!/bin/bash
 
-API_KEY="xxxx"
-API_EMAIL="xxx@gmail.com"
-ZONE_ID="xxx"
+#这些需要修改为自己的
+API_KEY="xxxxxxx"
+API_EMAIL="xxxx@gmail.com"
+ZONE_ID="xxxxxxxx"
 DOMAINS=("yx.example.com" "yx1.example.com" "yx2.example.com")
-pdir="/Users/mareep/Desktop" # 替换为你的路径
+pdir="/Users/mareep" # 替换为你的路径，如果此路径中存在CF/CloudflareST文件，则默认CloudflareST已经下载，如果没有则会自动下载解压
+urltest="https://urltest.example.com" #替换为自己的测速地址，托管在cf上的测速网站,自建的测速网站
+isIPv4=true #如果要测v6就改为false
+
+# 根据isIPv4的值执行不同的命令
+if [ "$isIPv4" = true ]; then
+    ipfile="ip.txt"
+    atype="A"
+    vtype="v4"
+else
+    ipfile="ipv6.txt"
+    atype="AAAA"
+    vtype="v6"
+fi
 
 arch=$(uname -m)
 case ${arch} in
@@ -20,24 +34,34 @@ case ${arch} in
         ;;
 esac
 
-cf_url="https://github.com/XIU2/CloudflareSpeedTest/releases/latest/download/CloudflareST_darwin_${cf_arch}.zip"
-curl -sL "$cf_url" -o "$pdir/cloudflaredST.zip"
-mkdir -p "$pdir/CF"
-unzip -q "$pdir/cloudflaredST.zip" -d "$pdir/CF"
-chmod +x "$pdir/CF/CloudflareST"
-echo ""
+
+if [ ! -f "$pdir/CF/CloudflareST" ]; then
+    echo "CloudflareST 不存在，开始下载，如果下载不动，手动下载解压到$pdir/CF/文件夹下."
+    cf_url="https://github.com/XIU2/CloudflareSpeedTest/releases/latest/download/CloudflareST_darwin_${cf_arch}.zip"
+    curl -sL "$cf_url" -o "$pdir/cloudflaredST.zip"
+    mkdir -p "$pdir/CF"
+    unzip -q "$pdir/cloudflaredST.zip" -d "$pdir/CF"
+    chmod +x "$pdir/CF/CloudflareST"
+else
+    # 如果文件存在，则跳过这段代码
+    echo "CloudflareST 已经存在，开始测速."
+fi
+
 
 domain_count=${#DOMAINS[@]}
 
 cd "$pdir/CF"
-"$pdir/CF/CloudflareST" -p $domain_count -n 1000 -f "$pdir/CF/ip.txt"
+"$pdir/CF/CloudflareST" -p $domain_count -n 1000 -f "$pdir/CF/$ipfile" -url $urltest
 bestips=($(awk -F ',' 'NR>1 {print $1}' "$pdir/CF/result.csv" | head -n ${#DOMAINS[@]}))
 
-for domain in "${DOMAINS[@]}"; do
-    record_info=$(curl -s -X GET "https://api.cloudflare.com/client/v4/zones/$ZONE_ID/dns_records?type=A&name=$domain" \
+for ((i=0; i<${#DOMAINS[@]}; i++)); do
+    domain="${DOMAINS[$i]}"
+    bestip="${bestips[$i]}"
+    record_info=$(curl -s -X GET "https://api.cloudflare.com/client/v4/zones/$ZONE_ID/dns_records?type=$Atype&name=$domain" \
         -H "X-Auth-Email: $API_EMAIL" \
         -H "X-Auth-Key: $API_KEY" \
         -H "Content-Type: application/json")
+    #echo "$record_info"
     record_id=$(echo "$record_info" | awk -F '"id":"' '{print $2}' | awk -F '","' '{print $1}')
     
     if [ -z "$record_id" ]; then
@@ -48,7 +72,8 @@ for domain in "${DOMAINS[@]}"; do
             -H "X-Auth-Email: $API_EMAIL" \
             -H "X-Auth-Key: $API_KEY" \
             -H "Content-Type: application/json" \
-            --data '{"type":"A","name":"'"$domain"'","content":"'"$bestip"'","ttl":1,"proxied":false}')
+            --data '{"type":"'"$atype"'","name":"'"$domain"'","content":"'"$bestip"'","ttl":1,"proxied":false}')
         echo "域名 $domain 的DNS记录已更新为IP地址 $bestip"
+
     fi
 done
